Binary swapping algorithm
I had an experiment in which I needed to know which of seven items contributed to a success.  There were a number of runs (67 in my case), so there were 67 opportunities to swap the items in and out.  There was then a test at the end of each run.  I could then assess the effectiveness of each of the seven items by calculating the difference it made to the test scores.

The program used code from another project, so it is not "nice", but it allows the user to enter the number of items (n) and the number of trials (t).  It then displays a list of binary number with n digits which can be copied/pasted into a text file.  A side panel also display the number of times each item appears in the experiment list.